+++ nav2_smac_planner/include/nav2_smac_planner/analytic_expansion.hpp
@@ -35,7 +35,7 @@ class AnalyticExpansion
 public:
   typedef NodeT * NodePtr;
   typedef typename NodeT::Coordinates Coordinates;
-  typedef std::function<bool (const unsigned int &, NodeT * &)> NodeGetter;
+  typedef std::function<bool (const uint64_t &, NodeT * &)> NodeGetter;
 
   /**
    * @struct nav2_smac_planner::AnalyticExpansion::AnalyticExpansionNodes
